+++ src/webots/maths/WbAxisAngle.hpp
@@ -1,0 +1,40 @@
+// Copyright 1996-2020 Cyberbotics Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef WB_AXIS_ANGLE_HPP
+#define WB_AXIS_ANGLE_HPP
+
+#include "WbVector3.hpp"
+
+//
+// Description: Axis-angle representation
+//
+
+class WbAxisAngle {
+public:
+  WbAxisAngle(double x, double y, double z, double angle) {
+    this->mAngle = angle;
+    this->mAxis = WbVector3(x, y, z);
+  };
+
+  WbVector3 &axis() { return mAxis; };
+
+  double angle() { return mAngle; }
+
+private:
+  WbVector3 mAxis;
+  double mAngle;
+};
+
+#endif
